requirements.txt
black==21.9b0
isort==5.9.3
mypy==0.910
pylint==2.11.1
PyYAML==5.4.1
